[!!/ED] Edit of first section
Edited bullets to have parallel structure and word order. Tersified language. Fixed oxford comma.

I suggested "video and audio media" in the first sentence to help tie the title "Accessible Media" to "video and audio"

Deleted "Transcripts are crucial for users who are deaf-blind, or for any user who relies on braille." as that belongs in earlier or later section, not this description bullet.

Consider if want to leave "Interactive transcripts can be synchronized with the media so that words are highlighted as they are spoken, and users can select words to move to that spot in the video." here in this short description, or not -- as it should be covered in the page with more details.

Overall I guess most of this is editor's discretion, yet *highly* recommended. :-)